requirements.txt
numpy 
opencv-python
mss 
pyautogui 
discord.py==2.3.2 
pywin32 
psutil
pytesseract